requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
tensorflow>=2.20.0
numpy>=1.26.0
Pillow>=10.0.0
opencv-python>=4.8.0
python-dotenv>=1.0.0
bcrypt>=4.0.1
Werkzeug>=2.3.7,<3.0
imagehash>=4.3.1
scikit-learn>=1.3.0
pandas>=2.0.3
matplotlib>=3.7.2
gunicorn>=21.2.0; platform_system != "Windows"
waitress>=3.0.0; platform_system == "Windows"
psycopg2-binary>=2.9.7